+++ templates/default/styles.css
@@ -1,5 +1,5 @@
 body {
-	background-color: rgb(240, 240, 240);
+
 	font-size: 12pt;
 	text-align: right;
 	color: #000;
