+++ templates/default/text_list.html
@@ -12,8 +12,8 @@
 </head>
 <body>
 	<p class="header_p">{{ title.title_id }} | {{ year_html|safe }}</p>
-	<div class="header">
-		{{ show_more_html|safe }}
-	</div>
+            <div class="header">
+                <a href="display_info/{{ title.title_id }}">{{ title.name_en }} ({{ title.name_ru }})</a>
+            </div>
 </body>
 </html>
